Refatoração do arquivo principal
- `Módulo app criado, onde vai estar a interface antiga`
- `Main alterada para chamar a função principal de app`
